moved forkchoice related tests to fork_choice_test.go
